refactor: S3ImageFixture 위치 common으로 이동
