+++ src/test/java/com/samsamotot/otboo/common/fixture/S3ImageFixture.java
@@ -1,4 +1,4 @@
-package com.samsamotot.otboo.common;
+package com.samsamotot.otboo.common.fixture;
 
 import org.springframework.mock.web.MockMultipartFile;
